VERY dirty almost working shaders
Im getting linker error for IGPUObjectsFromAssetConverter::getGPUObjectsFromAssets<ICPUSpeciliazedShader> and i have no idea why and not time to look into it more... so i couldnt test it. However this is some initial look of how this could seem as for now. This is very dirty and some things in the engine will not work properly:
* transparent scene nodes
* skinned scene nodes
* transform feedback
* BAW loading/writing
Also needs to be merged with devsh's cleanup branch (shaders in SMaterial should be smart_refctd_ptr, now they're just pointers)