Deconvolute the irrlichtbaw tutorial...
Really @Anastazluk the whole example should have been made without converting any `ICPU` objects via `getGPUObjectsFromAssets` except for `ICPUImage` and `ICPUBuffer`, should be making the GPU objects directly by hand whenever possible.

Furthermore its hard to explain what is the reason for usage of `SBasicViewParameters` and why it appears, it could lead someone to think every shader is like that (hint: it isn't) so I would create the GPU shaders MANUALLY from GLSL code stored in a file (+ a common header include with the `#ifdef __cplusplus` trick, and not even use the built-in `SBasicViewParameters` to demonstrate clarity) so that we can show the connection between C++ and GLSL structs.