We've unified instancing, boning and skinning we can run GPU boning at any instance count.
As for AABBs we will cache transformed AABBs with a GPU GLSL hashmap in the step between LoD selection and Drawcall culling via prefix sums and distributing one invocation to process one joint of a visible drawcall instance's skin. But thats in the far future.

For now we just make an overly conservative AABB for the animated mesh (containing the mesh during all animations.